+++ docs/md/releasenotes.md
@@ -8,6 +8,7 @@
 
 ### Bug fixes
 
+- Improve Kotlin MultiFileFacade metadata assertions to detect uninitialized references. 
 - Fix handling of category 2 values in `JvmValueTransferRelation` to work correctly with `ExecutingInvocationUnit`.
 - Fix concurrency problems in CallGraph and ValueFactory ids